requirements_for_pythonanywhere_virtenv.txt
python-dotenv~=0.17.0

Flask~=1.1.2

flask-bootstrap~=3.3.7.1
sqlalchemy-utils~=0.37.0

DateTime~=4.3
MarkupSafe~=1.1.1
numpy~=1.19.5

WTForms~=2.3.3
SQLAlchemy~=1.4.11

mysqlclient~=2.0.3
PyMySQL==1.0.2
pandas~=1.1.5
regex==2021.4.4
requests~=2.25.1
natsort>=7.1.1
Pillow>=8.1.1

flask_wtf~=0.14.3

Flask-SQLAlchemy>=2.5.1

email-validator==1.1.2
mysql-connector-python~=8.0.24 # for virtenv in pyhtnanwr, instead of mysqlclient~=2.0.3

docker>=5.0.3
